Add terminateCircularRelationships option (#31)
This introduces an option, `terminateCircularRelationships`, that when enabled, allows a particular type to only be resolved once by the mocks in a particular call stack. After the first time, subsequent resolutions will return an empty object.
This does not affect overrides, so if you need a circular structure, it can still be achieved using overrides. 

Fixes #13