fix(enumValues): update behaviour to match typescript plugin
There was some differences with typescript plugin when we were generating enum values.
For example `_id` was transformed to `Id` by typescript plugin, but we were generating `_Id`

We now use the same logic as typescript plugin, and changed the default to use `change-case-all#pascalCase` as the `change-case-all` package supports a lot of case converter.
We also use dynamic module resolution so you can pass anything in `enumValues` as long as it's compatible with `module#function` syntax