Add generatorMode to Options type, pass an appropriate value per node type
So, when we have a normal FieldDefinition, we want to assume we want to
use the output generator. When we visit the InputObjectTypeDefinition
and call generateMockValue on its children, we can assume we want the
input generator. We keep the other types of generator config types
around for backwards compatibility.

It could be a nice refactoring opportunity to map the existing config
types to the input/output type, instead of what we do in this commit,
which is check/switch on `value` in `getGeneratorDefinition`, but I
wanted to prove to myself that we could generate distinct mocks.