null-check value in getGeneratorDefinition
Tests were failing. Also, assert the last return is a GeneratorDefinition.

I have to clean this up, but essentially TS is being unable to narrow
this down. An InputOutputGeneratorOptions is a one level deep object
with GeneratorOptions, so the recursive call will inevitably trickle
down to a base case where we _do not_ have an InputOutputGeneratorOptions.